Move to array.rs
